Design a command line application or GUI that can take user input to select one of three modes of operation for the system, which are described below. This application will be run on the Pi via SSH and as such should be a part of the assignment3-pi branch.
Accept a user defined (x,y) location, go to that location, play a song, and return to base along any route. The (x,y) location is created by the user when selecting this mode.
Accept a user defined route of (x, y) locations, follow it, and return to base along any route. The route is created by the user when selecting this mode.
Explore with a wandering algorithm for a set amount of time. After the time is up, play a song, then return to base along any route. The amount of time is provided by the user when selecting this mode.

The map for this assignment is based on the size of the oLED with a 1” per pixel scale. Since the oLED is 128 x 64 pixels, the map area is 128” by 64”.

When accepting a mode of operation, the application will send a command to the MBED to display the mode along with any relevant information. This involves code in both of the branches of this assignment.

Display “Mode 1: Location” then change screens to display the (x,y) location.
Display “Mode 2: Route” then change screens to display the route.
Display “Mode 3: Explore” and the remaining time to explore.