During execution of all modes, the oLED will display the radar sweep visualization from Assignment 2. This will be managed in the assignment3-mbed branch.
During execution of all modes, the radar output need to be received and managed on the Pi in the form of a 128 x 64 occupancy grid.

Once a mode is complete and the iRobot has returned to base the MBED should stop radar output to the oLED. The Pi will provide to the MBED a description of the map state that it built up, which the MBED will display on the oLED. If a point on the occupancy grid is occupied, that pixel should be turned on.