The assignment4-remote branch will contain the application for the mbed-based remote control.
Use the LSM303 accelerometer board to sense rotational, forward and backward motion of the remote control and issue corresponding turn and drive commands to your iRobot.

Setup one button on the mbed to act as a deadman switch for control. This means if the user moves the controller around without pressing this button, nothing will happen. If the button is pressed, then control messages will actually be sent to the iRobot.

All messages produced will be sent over serial to a ZigBee radio to connect wirelessly to the iRobot.

Extra Credit: Set up a second button to trigger the iRobot to play a song.

Extra Credit: The LEDs on the mbed should all illuminate whenever the iRobot observes obstacles and is stopping to avoid it.